Collapse logic into one script
