--- maas_status_enum.py
@@ -1,2 +1,0 @@
-# @todo fill other states
-DEPLOYED = 'Deployed'
